+++ packages/extension/src/runner.ts
@@ -231,7 +231,6 @@ export function createRunnerManager({
             const { tables, routines } = unwrap(getChildrenResult);
 
             {
-              console.log(tables);
               const result = await renderer.renderTables(tables);
               if (!result.success) {
                 logger.error(result);